refactor: landing page and components
